+++ src/components/Select/CHANGELOG.md
@@ -14,3 +14,6 @@
 - 2025-10-31: Add the check mark for the selected item in the list.
 - 2025-11-04: Wrap the list with `ScrollArea` to make it scrollable when there
   are many options.
+- 2025-11-21: Add `width` prop to control the width of the component. Remove the
+  `min-width` constraint on the trigger. And remove the dependency on
+  `ThemeProvider`.